code/tiktok/tiktok_crawler/requirements.txt
uvicorn
fastapi
aiomysql
cryptography
python-json-logger
aiohttp
pydantic-settings
